First
I learn how to make an Air Canvas that can draw anything by simply catching the action of a colored marker with a camera. The marker in this case is a colored item near the tip of the finger. To develop this project, I'm utilizing OpenCV's computer vision techniques. Python is the ideal language due to its extensive libraries and simple syntax, however once the principles are understood, it can be implemented in any OpenCV supported language.